:truck: api: move FinalizedSession model from lib to src
