+++ api/src/certification/shared/domain/models/FinalizedSession.js
