+++ api/tests/certification/session-management/unit/infrastructure/serializers/to-be-published-session-serializer_test.js
@@ -1,5 +1,5 @@
-import { FinalizedSession } from '../../../../../../lib/domain/models/FinalizedSession.js';
 import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/to-be-published-session-serializer.js';
+import { FinalizedSession } from '../../../../../../src/certification/shared/domain/models/FinalizedSession.js';
 import { expect } from '../../../../../test-helper.js';
 
 describe('Certification | Session-management | Unit | Infrastructure | Serializers | to-be-published-session-serializer', function () {
