+++ api/tests/unit/domain/models/FinalizedSession_test.js
@@ -1,6 +1,6 @@
-import { FinalizedSession } from '../../../../lib/domain/models/FinalizedSession.js';
 import { JuryCertificationSummary } from '../../../../src/certification/session-management/domain/read-models/JuryCertificationSummary.js';
 import { CertificationIssueReportCategory } from '../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
+import { FinalizedSession } from '../../../../src/certification/shared/domain/models/FinalizedSession.js';
 import { status as assessmentResultStatuses } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import { domainBuilder, expect } from '../../../test-helper.js';
