+++ api/tests/unit/infrastructure/serializers/jsonapi/with-required-action-session_test.js
@@ -1,5 +1,5 @@
-import { FinalizedSession } from '../../../../../lib/domain/models/FinalizedSession.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/with-required-action-session-serializer.js';
+import { FinalizedSession } from '../../../../../src/certification/shared/domain/models/FinalizedSession.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | with-required-action-session-serializer', function () {
